requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
appnope==0.1.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.3
cssselect==1.2.0
decorator==5.1.1
exceptiongroup==1.0.4
execnet==1.9.0
fasteners==0.18
filelock==3.8.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
more-itertools==9.0.0
nose==1.3.7
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.32
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyotp==2.7.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-bdd==6.1.1
pytest-forked==1.4.0
pytest-html==2.0.1
pytest-metadata==2.0.4
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==3.0.2
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sbvirtualdisplay==1.1.0
selenium==4.6.0
seleniumbase==4.8.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.12
wcwidth==0.2.5
websockets==10.4
wsproto==1.2.0
